requirements.txt
starlette
uvicorn
fastmcp>=2.0.0
SQLAlchemy[asyncio]
asyncpg
python-dotenv
psycopg2-binary
Flask
gunicorn
flask-login==0.6.2
werkzeug==2.2.3
httpx
openai
flask-cors
keyring
